Moves handlers to own package
